# src/docs/swagger.yaml
definitions:
...
  dtos.CreateUserDto:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        minLength: 6
        type: string
      userName:
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    - userName
    type: object
  dtos.FailedResponseDto:
    properties:
      error:
        type: string
      errorType:
        type: string
      statusCode:
        type: integer
      statusText:
        type: string
    type: object
...
  dtos.SuccessResponseDto:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
      statusText:
        type: string
    type: object
...
  dtos.UpdateUserDto:
    properties:
      firstName:
        type: string
      lastName:
        type: string
    required:
    - firstName
    - lastName
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastLogin:
        type: string
      lastName:
        type: string
      password:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      userName:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: get all users
      produces:
      - application/json
      responses:
        "200":
          description: all users returned
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponseDto'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "400":
          description: token not passed with request
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "401":
          description: invalid/expired token
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "500":
          description: unexpected internal server error
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
      security:
      - JWT: []
      summary: returns all users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: create user
      parameters:
      - description: New User Details JSON
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: user created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponseDto'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: request body validation error
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "409":
          description: another user with supplied email exists
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "500":
          description: unexpected internal server error
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
      summary: registers a new user
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: delete user
      parameters:
      - description: User ID(UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: user deleted suuceesfully
          schema:
            $ref: '#/definitions/dtos.SuccessResponseDto'
        "400":
          description: request param validation error or token not passed with request
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "401":
          description: invalid/expired token
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "500":
          description: unexpected internal server error
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
      security:
      - JWT: []
      summary: deletes a user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: get user by ID
      parameters:
      - description: User ID(UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponseDto'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: request param validation error or token not passed with request
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "401":
          description: invalid/expired token
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "404":
          description: user with the specified ID not found
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "500":
          description: unexpected internal server error
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
      security:
      - JWT: []
      summary: returns a user by its 16 caharcter uuid
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: update user
      parameters:
      - description: User ID(UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User Details JSON
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: user updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dtos.SuccessResponseDto'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: request body/param validation error or token not passed with
            request
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "401":
          description: invalid/expired token
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "404":
          description: user with specified ID not found
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
        "500":
          description: unexpected internal server error
          schema:
            $ref: '#/definitions/dtos.FailedResponseDto'
      security:
      - JWT: []
      summary: updates a user
      tags:
      - User
swagger: "2.0"